// quantity-limiter-backend/src/modules/wix/types/wix.interface.ts
export interface IAppInstallEventPayload {
  refreshToken: string;
  appInstance: IAppInstanceResponse;
  siteData: ISiteProperties;
  shop: string;
  plan?: PricingPlan;
  dateLocale?: string;
}
// ...
export interface IWixQueryResponse {
  metadata: Metadata;
  totalResults: number;
}

export interface IWixProductResponse extends IWixQueryResponse {
  products: IWixProduct[];
}

export interface IWixCollectionResponse extends IWixQueryResponse {
  collections: IWixCollection[];
}

export interface IWixProduct {
  id: string;
  name: string;
  slug: string;
  visible: boolean;
  productType: string;
  description: string;
  stock: Stock;
  price: IWixPrice;
  priceData: IWixPrice;
  convertedPriceData: IWixPrice;
  additionalInfoSections: AdditionalInfoSection[];
  ribbons: Ribbon[];
  ribbon: string;
  brand: string;
  media: IWixMedia;
  customTextFields: CustomTextField[];
  manageVariants: boolean;
  productOptions: ProductOption[];
  productPageUrl: ProductPageUrl;
  numericId: string;
  inventoryItemId: string;
  discount: Discount;
  collectionIds: any[];
  variants: IWixProductVariant[];
  lastUpdated: string;
  seoData?: SeoData;
  sku?: string;
  weight?: number;
  createdDate: string;
  updatedDate: string;
}

export interface IWixCollection {
  id: string;
  name: string;
  media: IWixMedia;
  numberOfProducts: number;
  description: string;
  slug: string;
  visible: boolean;
}

export interface Stock {
  trackInventory: boolean;
  quantity?: number;
  inStock: boolean;
}

export interface IWixPrice {
  currency: string;
  price: number;
  discountedPrice: number;
  formatted: IWixFormatted;
}

export interface IWixFormatted {
  price: string;
  discountedPrice: string;
}

export interface AdditionalInfoSection {
  title: string;
  description: string;
}

export interface Ribbon {
  text: string;
}

export interface IWixMedia {
  mainMedia?: IWixItem;
  items: IWixItem[];
}

export interface IWixItem {
  thumbnail: IWixImage;
  mediaType: string;
  title: string;
  image: IWixImage;
  id: string;
}

export interface CustomTextField {
  title: string;
  maxLength: number;
  mandatory: boolean;
}

export interface ProductOption {
  optionType: string;
  name: string;
  choices: Choice[];
}

export interface Choice {
  value: string;
  description: string;
  media: IWixMedia;
  inStock: boolean;
  visible: boolean;
}

export interface ProductPageUrl {
  base: string;
  path: string;
}

export interface Discount {
  type: string;
  value: number;
}

export interface IWixProductVariant {
  id: string;
  choices: Choices;
  variant: IWixVariant;
}

export interface Choices {
  Size?: string;
  Color?: string;
}

export interface IWixVariant {
  priceData: IWixPrice;
  convertedPriceData: IWixPrice;
  weight: number;
  sku: string;
  visible: boolean;
}

export interface SeoData {
  tags: Tag[];
}

export interface Tag {
  type: string;
  children: string;
  custom: boolean;
  disabled: boolean;
  props?: IWixProps;
}

export interface IWixProps {
  name: string;
  content: string;
}

export interface Metadata {
  items: number;
  offset: number;
}